+++ src/main/java/com/informationsystem/library/model/Status.java
@@ -5,11 +5,11 @@ import static java.util.Map.entry;
 
 public class Status {
 
-    public static final Map<StatusName, Integer> STATUSES = Map.ofEntries(
-            entry(StatusName.IN_STOCK, 1),
-            entry(StatusName.CHECKED_OUT, 2),
-            entry(StatusName.TAKEN_OUT, 3),
-            entry(StatusName.LOST, 4)
+    public static final Map<StatusName, Short> STATUSES = Map.ofEntries(
+            entry(StatusName.IN_STOCK, (short)1),
+            entry(StatusName.CHECKED_OUT, (short)2),
+            entry(StatusName.TAKEN_OUT, (short)3),
+            entry(StatusName.LOST, (short)4)
     );
 
 }
